Implemented autocompleting dropdown for Category in add feedback
The form displaying category options has been updated to allow autocompletion.
The next step is to allow categories to be added directly from this dropdown

Also updated models to include a new Message model which is used for pre defined
messages for each category. This requires further improvement, but is acceptable at this
stage.

Issue #49
Ashwin
17/11/18