Implemented ability to choose pre defined message depending on category selected
The lecturer is now able to select a category, and then select from a list of messages associated
with that category.
This is slightly temperamental as before the category is chosen, the list of messages is incorrect.

Also need to implement ability to add new category and message in these sections.

Issue #49
Ashwin
17/11/18