Managed to create Category and Message directly in search bar
The function to allow a lecturer to either select from a list of categories or
create a new one has been implemented.
Similarly for the pre defined messages. A lecturer can now select a category, after which
the list of messages displayed will be based on that category. If no message exists or
they wish to add a new one, they can do this the same way as adding a new category.

As of now, the category field of the new Message made is not updated however, and so the next time
a lecturer tries to add feedback, messages will not appear for new categories.

Working on fixing this.

Issue #49
Ashwin
18/11/8