Students are now passed from 'lecturer_course' to 'add_group_feedback'
This is now done using cookies as this appeared to be the most suitable way to pass the
required data from one page to the next.
Now have to implemented functionality of actually adding the feedback for all students
that have been passed to the add_group_feedback page.

Issue #58