I have updated our program so that the course_token gets generated automatically in the models.
This is rather than it having to be done in the population script. This also fixed bugs that we were having when creating a class and the course_token wouldnt display on the webpage. So it now has full functionality. Tests are still to be written for it, as well as a button to be created on the courses to "create Course"