Have finished implementation of the create course button.
I have put this at the top of the card, so that if the lecturer has a long list of classes he won't have to scroll to the bottom to get to the button. I have also made it stretch the width of the card and centre aligned the button text.