Fixed bugs, and added leaderboard for when a student views a course
There was a bug that when a student viewed one of their courses, the feedback displayed
was actually all feedback given to that course. This has been fixed so that only the students
own feedback for that course is displayed.
The leaderboard for each course has also been implemented for the student.

Made updates to the model, adding a few new functions to keep the work done in the view
as simple as possible, and to improve the separation of concerns