Able to provide individual feedback for first selected student
A new "Add Individual Feedback" button has been added. A lecturer can select a number of
students and upon pressing this button, they will be redirected to the "add-individual-feedback"
page for the first selected student.
The other students are currently being stored as a cookie, so theoretically I should be able
to remove the current student from the cookie as soon as that page loads, and upon clicking
"Submit" redirect to the next student (if there is one) and repeat until no students
are left at which point redirect to "my-provided-feedback"