Now able to add a new message for a given category
Simply click on a category, and then add a message as required.
Also had to fix a bug which meant that adding feedback was not working (this was because
the Category model no longer has primary_key 'name' so the user parameter needs to be filled as
well)
Seems to be another bug with DAL creating pre defined message. Will try to fix this.

Issue #100