Can now provide feedback as a student.
This was temporarily removed to improve the way in which it was done. Now a student simply
navigates to a course and selects a student to give feedback to. If they try to give feedback
to themself then they get an error.

Just realised another bug that needs to be fixed - if the student searches the required url
to give themself feedback then they will be able to. Will make an issue to fix this.

Issue #132